Add a -c/--check option to help regenerate file
Run with -c / --check=path/to/allowed.yml to run _only_ the listed
allowed failures, and regenerate the file with _only_ the ones that are
still failing.

This option is very useful when you have a large test suite and a lot of
allowed failures, and you fix _one_ bug, instead of having to edit the
file by hand.